requirements.txt
mysql-connector-python
hydra-core
IPython
hydra-joblib-launcher
torch
numpy
black
gymnasium
rich
jaxtyping
hydra-submitit-launcher
